+++ config/autoload/preview-generation.global.php
@@ -1,6 +1,7 @@
 <?php
 declare(strict_types=1);
 
+/** @deprecated */
 return [
 
     'preview_generation' => [