+++ module/PreviewGenerator/src/Service/PreviewGeneratorInterface.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Common\Service;
+namespace Shlinkio\Shlink\PreviewGenerator\Service;
 
-use Shlinkio\Shlink\Common\Exception\PreviewGenerationException;
+use Shlinkio\Shlink\PreviewGenerator\Exception\PreviewGenerationException;
 
 /** @deprecated */
 interface PreviewGeneratorInterface