+++ module/Common/src/Rest/DataTransformerInterface.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Shlinkio\Shlink\Common\Rest;
-
-interface DataTransformerInterface
-{
-    public function transform($value): array;
-}