+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 * [#450](https://github.com/shlinkio/shlink/issues/450) Added PHP 7.4 to the build matrix, as an allowed-to-fail env.
 * [#443](https://github.com/shlinkio/shlink/issues/443) Split some logic into independent modules.
+* [#451](https://github.com/shlinkio/shlink/issues/451) Updated to infection 0.13.
 
 #### Deprecated